Added a generator for the  Watts-Strogatz model.
Also changed some remaining floats into doubles in the BaseGenerator.

Also updated the javadoc of the Generator interface that was saying
false things.

+++ src/org/graphstream/algorithm/generator/BaseGenerator.java
@@ -452,7 +452,7 @@ public abstract class BaseGenerator extends SourceBase implements Generator {
 	 * @param y
 	 *            The node ordinate.
 	 */
-	protected void addNode(String id, float x, float y) {
+	protected void addNode(String id, double x, double y) {
 		addNode(id);
 		sendNodeAttributeAdded(sourceId, id, "xy", new Double[] {
 				new Double(x), new Double(y) });
